requirements.txt
azure-storage-blob
sentence-transformers
faiss-cpu
python-dotenv